+++ src/lcmtypes/mbot_imu_t.hpp
@@ -70,13 +70,13 @@ class mbot_imu_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int mbot_imu_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = getHash();
+    int64_t hash = (int64_t)getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -180,9 +180,9 @@ int mbot_imu_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-int64_t mbot_imu_t::_computeHash(const __lcm_hash_ptr *)
+uint64_t mbot_imu_t::_computeHash(const __lcm_hash_ptr *)
 {
-    int64_t hash = 0x2e9f8f63a10a11c1LL;
+    uint64_t hash = 0x2e9f8f63a10a11c1LL;
     return (hash<<1) + ((hash>>63)&1);
 }
