+++ src/lcmtypes/robot_path_t.hpp
@@ -75,13 +75,13 @@ class robot_path_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int robot_path_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = getHash();
+    int64_t hash = (int64_t)getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -171,7 +171,7 @@ int robot_path_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-int64_t robot_path_t::_computeHash(const __lcm_hash_ptr *p)
+uint64_t robot_path_t::_computeHash(const __lcm_hash_ptr *p)
 {
     const __lcm_hash_ptr *fp;
     for(fp = p; fp != NULL; fp = fp->parent)
@@ -179,7 +179,7 @@ int64_t robot_path_t::_computeHash(const __lcm_hash_ptr *p)
             return 0;
     const __lcm_hash_ptr cp = { p, (void*)robot_path_t::getHash };
 
-    int64_t hash = 0xd8a57fd0b3392990LL +
+    uint64_t hash = 0xd8a57fd0b3392990LL +
          pose_xyt_t::_computeHash(&cp);
 
     return (hash<<1) + ((hash>>63)&1);
