+++ src/lcmtypes/wifi_data_t.hpp
@@ -67,13 +67,13 @@ class wifi_data_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int wifi_data_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = getHash();
+    int64_t hash = (int64_t)getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -173,9 +173,9 @@ int wifi_data_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-int64_t wifi_data_t::_computeHash(const __lcm_hash_ptr *)
+uint64_t wifi_data_t::_computeHash(const __lcm_hash_ptr *)
 {
-    int64_t hash = 0x24a240b71ffcb820LL;
+    uint64_t hash = 0x24a240b71ffcb820LL;
     return (hash<<1) + ((hash>>63)&1);
 }
